# swagger.yaml
openapi: 3.0.0
info:
  title: Marvel API
  description: Marvel Demo for xendit challenges
  version: 0.0.1
servers:
  - url: http://localhost:8080
paths:
  /characters:
    get:
      summary: Returns a list of character ids.
      parameters:
        - in: query
          name: offset
          schema: 
            type: integer
          required: false
        - in: query
          name: limit
          schema: 
            type: integer
          required: false
      responses:
        '200':    # status code
          description: Array of character ids
          content:
            application/json:
              schema: 
                type: object
                properties:
                  data:
                    type: array
                    items: 
                      type: integer
        '500':
          $ref: '#/components/responses/Error'

  /characters/{id}:
    get:
      summary: Return character based on id
      parameters:
        - in: path
          name: id
          schema: 
            type: integer
          required: true
      responses:
        '200':    # status code
          description: character detail
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    $ref: '#/components/schemas/Character'
        '500':
          $ref: '#/components/responses/Error'
components:
  responses:
    Error:
       description: Internal server error
       content:
          application/json:
            schema:
              type: object
              properties:
                error:
                  $ref: '#/components/schemas/Error'
  schemas:
    Character:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        description:
          type: string
      required:
        - id
        - name
    Error:
      type: object
      properties:
        code:
          type: string
        message:
          type: string
      required:
        - code
        - message